fix: Move DTOs to the /inbound/presenters folder
