+++ src/domain/use_cases/app/produto.use_case.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { ProdutoService } from 'src/adapters/outbound/repositories/produto.service';
 import { ProdutoEntity } from 'src/domain/entities/produto.entity';
-import { CriaProdutoDTO } from 'src/adapters/outbound/models/dto/CriaProduto.dto';
-import { AtualizaProdutoDTO } from 'src/adapters/outbound/models/dto/AtualizaProduto.dto';
+import { CriaProdutoDTO } from 'src/adapters/inbound/rest/v1/presenters/dto/CriaProduto.dto';
+import { AtualizaProdutoDTO } from 'src/adapters/inbound/rest/v1/presenters/dto/AtualizaProduto.dto';
 import { randomUUID } from 'crypto';
 
 @Injectable()
